backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
c>=1.7.4
python-multipart>=0.0.5
python-dotenv>=0.19.0
tavily-python>=0.1.2
qdrant-client>=1.1.1
langchain>=0.0.200
langchain-community>=0.0.10
llama-index>=0.9.0
groq>=0.4.0
sentence-transformers>=2.2.2
